Creating docker container
