+++ azure_edge_iot_ops_jumpstart/aio_esa/artifacts/esa_ai/requirements.txt
@@ -1,0 +1,13 @@
+Flask==3.0.2
+opencv-python==4.9.0.80
+numpy==1.24.4
+matplotlib>=3.3.0
+pillow>=7.1.2
+pyyaml>=5.3.1
+requests>=2.23.0
+scipy>=1.4.1
+tqdm>=4.64.0
+pandas>=1.1.4
+psutil
+dill
+py-cpuinfo
