--- azure_edge_iot_ops_jumpstart/aio_esa/modules/esa_ai/index.html
@@ -1,41 +1,0 @@
-<html lang="en" data-bs-theme="dark">
-<head>
-  <title>Jumpstart ESA Scenario</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" crossorigin="anonymous">
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz"
-    crossorigin="anonymous"></script>
-  <link href="https://cdn.datatables.net/v/bs5/jq-3.6.0/jszip-2.5.0/dt-1.13.5/af-2.6.0/b-2.4.0/b-colvis-2.4.0/b-html5-2.4.0/b-print-2.4.0/cr-1.7.0/date-1.5.0/fc-4.3.0/fh-3.4.0/kt-2.10.0/r-2.5.0/rg-1.4.0/rr-1.4.0/sc-2.2.0/sb-1.5.0/sp-2.2.0/sl-1.7.0/sr-1.3.0/datatables.min.css" rel="stylesheet" />
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.2.7/pdfmake.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/pdfmake/0.2.7/vfs_fonts.js"></script>
-  <script src="https://cdn.datatables.net/v/bs5/jq-3.6.0/jszip-2.5.0/dt-1.13.5/af-2.6.0/b-2.4.0/b-colvis-2.4.0/b-html5-2.4.0/b-print-2.4.0/cr-1.7.0/date-1.5.0/fc-4.3.0/fh-3.4.0/kt-2.10.0/r-2.5.0/rg-1.4.0/rr-1.4.0/sc-2.2.0/sb-1.5.0/sp-2.2.0/sl-1.7.0/sr-1.3.0/datatables.min.js"></script>
-</head>
-
-<body>
-  <div class="container">
-    <div class="row">
-      <img src="../static/jumpstart_scenarios.png" class="logo">
-    </div>
-    <div class="row">
-      <img src="{{ url_for('video_feed') }}" class="video-feed">
-    </div>
-    <div class="row" style="margin-top:50px; margin-left: 100px; margin-right: 100px;">
-      <table id="file-table" class="table table-sm table-responsive pb-3" width="100%">
-        <thead>
-          <tr>
-            <th>Name</th>
-            <th>Size (Kb)</th>
-            <th>Last Modified</th>
-          </tr>
-        </thead>
-        <tbody>
-        </tbody>
-      </table>  
-    </div>
-  </div>
-</body>
-<link rel="stylesheet" href="../static/index.css">
-<script src="../static/index.js"></script>
-
-</html>
