--- azure_edge_iot_ops_jumpstart/aio_esa/modules/esa_ai/requirements.txt
@@ -1,3 +1,0 @@
-Flask==1.1.2
-opencv-python==4.5.1.48
-ultralytics==0.0.6
