--- azure_edge_iot_ops_jumpstart/aio_esa/modules/esa_ai/static/index.css
@@ -1,15 +1,0 @@
-body {
-    background-color: black;
-}
-
-.logo {
-    width: 30%;
-    margin:auto;
-    display:block; 
-}
-
-.video-feed {
-    width: 60%;
-    margin:auto;
-    display:block; 
-}
